+++ WebStore/Infrastructure/Services/InMemory/InMemoryEmployeesDataService.cs
@@ -6,7 +6,7 @@ using WebStore.Data;
 using WebStore.Infrastructure.Interfaces;
 using WebStore.Models;
 
-namespace WebStore.Infrastructure.Services
+namespace WebStore.Infrastructure.Services.InMemory
 {
     public class InMemoryEmployeesDataService : IEmployeesDataService
     {
